+++ src/GeometricFlux.jl
@@ -13,13 +13,6 @@ using Flux: glorot_uniform, leakyrelu, GRUCell
 using Flux: @functor
 using ZygoteRules
 
-import Base: identity
-import Base.Threads: atomictypes, llvmtypes, inttype, ArithmeticTypes, FloatTypes,
-       atomic_cas!, atomic_xchg!,
-       atomic_add!, atomic_sub!, atomic_max!, atomic_min!,
-       atomic_and!, atomic_nand!, atomic_or!, atomic_xor!
-import Base.Sys: ARCH, WORD_SIZE
-
 export
 
     # layers/meta
