+++ src/atomic.jl
@@ -1,4 +1,10 @@
 # from https://github.com/chengchingwen/Transformers.jl/tree/master/src/fix
+import Base: identity
+import Base.Threads: atomictypes, llvmtypes, inttype, ArithmeticTypes, FloatTypes,
+       atomic_cas!, atomic_xchg!,
+       atomic_add!, atomic_sub!, atomic_max!, atomic_min!,
+       atomic_and!, atomic_nand!, atomic_or!, atomic_xor!
+import Base.Sys: ARCH, WORD_SIZE
 
 for typ ∈ atomictypes
     lt = llvmtypes[typ]
