+++ src/layers/pool.jl
@@ -1,13 +1,3 @@
-floattype(::Type{T}) where {T<:AbstractFloat} = T
-floattype(::Type{Int8}) = Float16
-floattype(::Type{UInt8}) = Float16
-floattype(::Type{Int16}) = Float16
-floattype(::Type{UInt16}) = Float16
-floattype(::Type{Int32}) = Float32
-floattype(::Type{UInt32}) = Float32
-floattype(::Type{Int64}) = Float64
-floattype(::Type{UInt64}) = Float64
-
 struct GlobalPool{A}
     aggr::Symbol
     cluster::A
