+++ src/utils.jl
@@ -1,3 +1,19 @@
+## Type transformation
+
+floattype(::Type{T}) where {T<:AbstractFloat} = T
+floattype(::Type{Int8}) = Float16
+floattype(::Type{UInt8}) = Float16
+floattype(::Type{Int16}) = Float16
+floattype(::Type{UInt16}) = Float16
+floattype(::Type{Int32}) = Float32
+floattype(::Type{UInt32}) = Float32
+floattype(::Type{Int64}) = Float64
+floattype(::Type{UInt64}) = Float64
+
+
+
+## Inverse operation of scatter
+
 function gather(input::AbstractArray{T,N}, index::AbstractArray{<:Integer,N}, dims::Integer;
                 out::AbstractArray{T,N}=similar(index, T)) where {T,N}
     @assert dims <= N "Specified dimensions must lower or equal to the rank of input matrix."
@@ -10,7 +26,6 @@ function gather(input::AbstractArray{T,N}, index::AbstractArray{<:Integer,N}, di
     return out
 end
 
-
 function gather(input::Matrix{T}, index::Array{Int}) where T
     out = Array{T}(undef, size(input,1), size(index)...)
     @inbounds for ind = CartesianIndices(index)
@@ -19,7 +34,19 @@ function gather(input::Matrix{T}, index::Array{Int}) where T
     return out
 end
 
+function gather_indices(X::Array{T}) where T
+    Y = DefaultDict{T,Vector{CartesianIndex}}(CartesianIndex[])
+    @inbounds for (ind, val) = pairs(X)
+        push!(Y[val], ind)
+    end
+    Y
+end
+
 identity(; kwargs...) = kwargs.data
+
+
+
+## Graph related utility functions
 
 struct GraphInfo{A,T<:Integer}
     adj::AbstractVector{A}
@@ -41,6 +68,8 @@ function edge_index_table(adj::AbstractVector{<:AbstractVector{<:Integer}},
     y .= 0, cumsum(map(length, adj))...
     y
 end
+
+
 
 ## Indexing
